Update patch set 5
Patch Set 5:

[ +fdoray, who I believe wrote this originally ]

Thanks for contributing this! I don't really know this code and so I don't have a lot of context on either it or the linked bug, so it might be good if fdoray@ or mmoroz@ could take a look at it.

At first glance, though, my thought is that it's not clear that splitting based on the size of the CL is the right way to look at it. Generally speaking, we try to group CLs based on some sort of semantic meaning, not just size. The original version of split would group by OWNERS, which is reasonable (why show an owner files they don't care about), and the bug filed by mmoroz@ I think was to try and cluster by higher-level owners, to reduce the total number of CLs.

Neither of these things involves splitting by size, as a result, so I'm not sure if this is an improvement.

On the other hand, I've also never used this, so maybe splitting by size *would* be an improvement.

Separately, I don't know why you're using a trie for this, or this third-party trie in particular. Could you explain that a bit more?

Patch-set: 5
Reviewer: Gerrit User 1002897 <1002897@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1149974 <1149974@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>