Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > Will this work equally well with MinGW/cmd/Cygwin?  It looks like some of the comments refer to "Windows" but only check MinGW, or the like.
> 
> It should work as well as it used to.

I guess I'm mostly saying it's not clear from the comments whether the omission was intentional or an oversight.  Maybe if MinGW should be handled specially and the comment says "MinGW" instead of "Windows" no one will be tempted in the future to add a cmd case because they mistakenly think it was forgotten, or something.

Anyway, not a reviewer here, so you're welcome to ignore me.

Patch-set: 6