Update patch set 5
Patch Set 5:

maybe you could explain the CQ failures earlier in here ?

the first time i tried to submit, i didn't have the footer in the commit message:
https://ci.chromium.org/b/8901955256863576960

so i added the footer and tried to submit again.  but that failed because it said the footer shouldn't be there:
https://ci.chromium.org/b/8901954701497901056

so i took the footer back out and tried to submit a third time.  that time it worked.

Patch-set: 5
CC: Gerrit User 1001629 <1001629@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>