git-cache: don't limit pack files to 2GiB.
This is only relevant to 1 builder which updates git cache bundles.
Since some time in 2018-2019, it started creating several 2GiB packs
for chromium/src. I verified that the setting affecting it is
  core.deltaBaseCacheLimit

Log file of a build which I observed while ssd-ed into the machine (internal):
  https://logs.chromium.org/logs/infra-internal/buildbucket/cr-buildbucket.appspot.com/8901682638839779824/+/steps/Updating_https:__chromium.googlesource.com_chromium_src/0/stdout
Git version: version:2.21.0.chromium16

R=ehmaldonado

Change-Id: I52cadfb9f34faea09a57d53387cab7e0538362b9