Remove unused private functions in git_cl
There are many functions in git_cl that appear unused, although
it's possible that some of them are used in other places, where
git_cl may be imported as a library.

This CL just removes some of the private ones, which at least
were not supposed to be used from other places.

Change-Id: Ib8eee0dfb181c8cda04c0894d2daea52b3ad42ac