Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 2:
> 
> PTAL
> 
> This shouldn't break anything, since hooks are already executed on vpython on bots.

Unfortunately, that's not the whole world. I can easily see this breaking things in dev checkouts and/or on other projects or other bots.

That said, I'm okay with landing this and finding out. I suspect that if it *does* break things we might need some way to manage the rollout via a flag in DEPS files to opt-in or something.

Patch-set: 3
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1