Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i grok that this is only addressing gclient.  in my ideal, we'd set the shebangs for all our scripts to the theoretical "vpython2" so they'd all be forced to python 2.  only once scripts/projects had been updated to support python 3 would they use "vpython" or "vpython3".

That's an ideal for me too. There are a couple bugs blocking this, but we're actively working on it. Using vpython would be enough if we add a .vpython file asking for a specific version of Python.

I also plan to add something like gclient3, or maybe an environment variable (though that is a bit inconvenient to set) so users can explicitly try python3.

In the meantime, I'd like to make this a warning instead of an error, and skip the python3_migration env var part.

Patch-set: 1
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>