Make gclient.py require Python 2
An increasing number of Windows developers are "accidentally" ending up
with Python 3 installed on their machine through the Windows Store. If
this version is first in the path then gclient may fail in non-obvious
ways. This adds a python version check to give a clearer error message.

If the python3_migration environment variable is set, to anything, then
the check will be ignored. This allows work on Python 3 compatibility to
continue.

R=vapier@chromium.org

Change-Id: I7a292fba2ab7d059a6149644d0b1fd92d2371dce