Update patch set 4
Patch Set 4:

Hi Andrii,
I wrote this script for working with worktrees and the chromium repo, and I've found it pretty useful, and want to share it w/ some of my teammates.  This seemed to me like a reasonable place to check it in.  LMK what you think.
Danan

Patch-set: 4
Work-in-progress: false