Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > The error message aside, can we accommodate setups where having gclient automatically (re)start goma is not desirable?
> 
> I think it would be awkward to update binaries but not restart.
> 
> Can you make things work by just disabling auto-update? That works today by setting DEPOT_TOOLS_UPDATE=1 in your environment and/or creating a .disable_auto_update file in the repo dir.

That would mean inevitably in a couple of weeks I'm unable to build chromium when there's a different update that's required by a chromium change. Everyone needs (current) depot tools, not everyone wants or can use goma ,and specifically not everyone will want the binaries from depot_tools. While I can ignore and simply not use the binaries, it's a lot harder to deal with an auto-starting daemon.

Patch-set: 3