chromite_wrapper: modernize a bit
This file hasn't been updated in over 6 years.  The backwards compat
logic in here dates back to pre-R20 releases, and the only boards we
released that long ago are EOL now.  So scrub all the unused code to
simplify it.

While we're here, fix a few style issues, and update the docs.

Bug: 997354
Test: `cros help` still works