Update patch set 4
Patch Set 4:

Are we sure that updating the binary while the server is in use won't cause problems if the server isn't restarted? Do we actually update the binary in-place, or does the binary get moved to a different path thanks to the way cipd manages it?

(I think we discussed this during the design of this, but I've forgotten the answer :( ).

Patch-set: 4
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>