Update patch set 4
Patch Set 4:

Is there a way we can detect that goma is running and at least warn the user?

And/or can we detect that the goma that is running is indeed the binary we just updated and that it's safe to update it?

Patch-set: 4
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>