Allow the bot_update api to not fetch tags from the git server.
There are 3 layers modified, starting from the bottom up:

1. git_cache.py populate

Now takes a --fetch-tags option. If specified, the cache will fetch
updated tags from the server. If not, all fetches will be done with
--no-tags, which prevents the server from sending all the tag refs. In
chromium this prevents significant time bottlenecks dealing with 10k+
tags.

2. bot_update.py

bot_update has to deal with multiple git repos, it has the root repo
that is checked out through git-cache, as well as repos synched via
DEPS and gclient.

The script now takes a --fetch_checkout_tags option. If specified,
the git-cache populate call will include --fetch-tags. Otherwise, it
won't. So this controls (for chromium) if fetches to the src.git server
are passed --no-tags.

3. bot_update/api.py

This is the entry point for recipes and also has to deal with multiple
git repos. The behaviour at this point is not changed from the default.
A |fetch_checkout_tags| parameter is added to ensure_checkout() that
defaults to true.

This changes the test expectations as all bot_update recipes now include
a --fetch_checkout_tags when they call through to bot_update.py.


This CL is a refactor with no intended behavior change.

The next step will be to change the chromium build repo to pass False
for ensure_checkout(fetch_checkout_tags) on chromium trybots.

This represents solution #2 in
https://docs.google.com/document/d/1hDIunJjjfpmr50y3YnZ4o3aq1NZF4gJa1TS0p7AIL64/edit#

Bug: 1019824
Change-Id: I935430603299daa9e301a95a5184c0ce486fd912