Use git fetch --no-tags in gclient sync if --with_tags not specified
When --with_tags is specified, we add refs/tags/* to the fetch specs
in gclient. Otherwise, we fall back to default behaviour which fetches
tags that point to commits that are being fetched otherwise.

This creates slow bottlenecks if a repo such as chromium has 10k+ tags.
So when --with_tags is not specified, we can use the other non-default
behaviour and not fetch ant tags. This prevents the server from sending
the list of all tags and the client from processing them.

R=tandrii@chromium.org

Bug: 1019824
Change-Id: I756beb8c2b0c7ec3c0fc48b7431d0cf8c0bbbbd3