Change "CPU time" to "elapsed time"
The "CPU time" values printed by post_build_ninja_summary.py are not
actually measure of CPU time, they are measures of wall-clock time or
elapsed time. This just changes the labeling of the numbers - the actual
numbers printed are unchanged.

The "weight time" number continues to be the more important number since
it is an approximation of how much the build step(s) are affecting build
time. See the comments in the script for details.

Change-Id: Ibdb4efdd327ece34492ab10337c234a826514197