depot_tools: Execute tests using bot's vpython instead of depot_tools' vpython
On Windows, scripts on the current directory take precedence over PATH, so
that when testing depot_tools on Windows, calling `vpython.bat` will
execute the copy of vpython of the depot_tools under test instead of the
one in the bot.
As a workaround, we run the tests from the parent directory instead.

Change-Id: I5187943b6cf7499d50b3f1a2ec906a5e8fd84835