Update patch set 7
Patch Set 7:

I played around with the tests for a bit and couldn't really get them to meaningful test anything about this, because the change is about cleaning up input before it gets to the internals. The real tests are the asserts which validate that backslashes don't get into the data structures so that the code doesn't need to handle them.

I'm going to keep using this for a bit to make sure all of the inputs are sanitized before landing. Let me know if you have any ongoing concerns.

Patch-set: 7
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>