Update patch set 1
Patch Set 1: Commit-Queue+1

The goal is to support `include ` as a drop-in alternative to `file:`. 

I understand that there might be other longer-term efforts to update this, but in the meantime, this appears like a low-hanging fruit.


This change helps better align the git-cl owners syntax with the syntax described in Gerrit's find-owners plugin:
https://gerrit.googlesource.com/plugins/find-owners/+/master/src/main/resources/Documentation/syntax.md


Note that the find-owners include syntax offers the additional ability to specify cross-repo files, but that is not being supported here.

Patch-set: 1
Reviewer: Gerrit User 1354087 <1354087@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Work-in-progress: false