Update patch set 9
Patch Set 9:

I need advice on how to handle this. I got a couple of teammates to use the patched version for a while and one of them hit the assert during "git cl upload". The function that needs the sanitization now (the function that calls in to owners_py) is _CheckAddedDepsHaveTargetApprovals in Chromium's PRESUBMIT.PY. This is worrisome because I don't want this fix to break other projects or require multi-depot changes.

I tried centralizing the sanitization in files_not_covered_by but that is messy because it is normally passed a list for "files" but in the unit tests is passed a string, set, and generator.

Meanwhile the asserts haven't actually found any bugs, so I don't want to complicate the code too much in order to support them.

The current version has some very messy changes in files_not_covered_by (too messy to submit as-is) and still hits the asserts during "git cl upload" (in a call from _CommonChecks)

So, any great ideas? Removing the asserts and simplifying the change is all I can think of.

Patch-set: 9
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>