[roll] Fix the command in the commit message
The CL https://crrev.com/c/chromium/tools/depot_tools/+/2031104 used
the full hash when calling "git log" to fix the roll on some bot. It
however resulted in the full hash visible in the commit message.

Fix the command included in the commit message while still using the
full hash in the command invoked.

Bug: none
Change-Id: I454e600fc34e554c1186ae9488ea4b0f7e7df310