Update patch set 1
Patch Set 1:

> > I don't have my machine at hand now, but I did almost all the packaging work on Windows Server 2019 Datacenter Edition so there are good chances that it is the same.
> > 
> > I would wait for Bruce to be back and take a look at this.
> 
> OK, going to wait for Bruce.

Back when we hit this problem initially the file would be randomly different. As in, I'd create two identical VMs and install VS and the contents of windows.winmd would be different. It sounds like they might have fixed that issue.

Even if they haven't fixed it, if we need the file then we need it, and we should package it. If other companies can't recreate our toolchain hash then we have a mechanism for them to redirect to their hash so they should be fine, I think.

Patch-set: 1
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>