Update patch set 2
Patch Set 2:

> Patch Set 2: -Auto-Submit
> 
> i went with vpython3 so i wouldn't have to worry about people having old crufty versions of python3.  e.g. the goma people trying to build on ancient trusty w/python-3.4.

Ah... so this is a common misconception with vpython; it doesn't actually pull a new interpreter, it walks $PATH to find an interpreter which self-reports a version that's compatible with the spec (so in this case, 3.8.x). If it fails to find one, it will die with a message to that effect. That _might_ be good-enough for your purposes here (it would at least prevent getting halfway into things before an incompatibility is discovered)

> 
> i don't think the one-time hit with pulling the default vpython3 spec is a big deal here.  it should be amortized in general for devs right ?

Yes it's cached (including on bots).

Patch-set: 2
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>