Fix build summary to classify *.mojom*.obj as .obj
The build summary script was overly aggressive about summarizing .mojom
files under the mojo category. This was intended to measure the cost of
generating source from mojo files but it accidentally pulled in the
.mojom object files.

Categorizing different types of object files might be worthwhile but
that should be done intentionally and consistently.

Change-Id: Iab6b7e94797ce7f1ed46805034b4f274c88617e0