Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> It looks this would get reverted once python3 is only supported version. Why not update `presubmit_support.py` and `PRESUBMIT.py` to be compatible with python3 first and avoid subprocess calls (ie this cl)? Note that presubmit_support.py already had some logic to handle py2/py3, at least in import statements.

presubmit_support.py is already compatible with Python 3.
The problem is that we don't own the PRESUBMIT.py scripts, and there are many of them in several repos, so updating all of them is not a small project.
So instead of tackling that now, we can decouple git-cl and PRESUBMIT.py scripts, so we can update git-cl to Python 3, and tackle the PRESUBMIT.py migration later.

Patch-set: 8
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0