Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Thanks for the heads-up. A niche feature that probably won't be discovered but I imagine it could be handful to examine all *mojo* files for instance. If the target names include a path it could be useful for checking various folders too, but again, I wouldn't expect this to be discovered by anyone. Hopefully curious people ask around.

I used an early version of it to work on this bug - crbug.com/1054626. mojo costs are indeed getting high, and this can help reveal it. Discoverability is a problem, but only a few people need to know about this (and I'll mention it in an upcoming blog post)

Patch-set: 4
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0