[man] Fix order of depot_tools in PATH
For development in V8, and I guess also in Chrome, it is important that
depot_tools are at the beginning of $PATH. This was explicitly mentioned
in the original tutorial. However, it seems like a cleanup CL changed
this recently, see https://crrev.com/c/1593820. This CL changes the
order in $PATH back to the original order.

CC=marja@chromium.org, dmcardle@chromium.org
R=iannucci@chromium.org

Change-Id: Idaed3ada05cddd666734c03a25503fe5dc481e3a