Tweak git bat template
This fixes some issues with escaping, namely carets. For example,
without this patch, when HEAD^^1 is passed to batch script, it will
forward HEAD1 to git. With this patch, it will be forwarded as HEAD^^1
to git. There is still issue with HEAD^1 which seems not possible to fix
with batch script as batch receives HEAD1.

R=ehmaldonado@chromium.org

Change-Id: Ibb48ef06b4f17df374ee983eaa19ca43a706fa22
Bug: 1065307