Do not prompt on --dependencies if --force.
At the moment, specifying --dependencies to upload all dependent CLs will always
prompt for confirmation if --force was specified. This means that there is no
way to run this command in a batch way or from another script, and it's also in
contradiction to the --force flag which is meant to answer 'yes' to all prompts.

Change-Id: I7d85b02b68d958a5b2d3295a04c4ced023d36ec8