Propagate source manifest fetch_ref
Angle team needs Milo to display blamelists for multiple repos:
crbug.com/1047893.
For that Milo, needs to know which repo in the source manifest [1] was
fetched from a git ref, as opposed to pinned to a specific commit.
Manifest has fetch_ref field for that, but bot_update currently does not
populate it.

Set the field to refs/heads/master if the requested revision is "HEAD"
or "master". Otherwise, set the field to the requested revision
if it looks like a git ref (starts with "refs/). None of this
is as precise as we'd want, but is consistent with bot_update's current
level of precision/correctness.

[1]: https://source.chromium.org/chromium/infra/infra/+/master:go/src/go.chromium.org/luci/common/proto/srcman/manifest.proto;l=32

Bug: 1047893
Change-Id: Ic8dca388e3e5a60a5a6743f66c56e78082488256