cpplint: move a build/include check to subcategory
There is a cpplint include check that ensures that all includes have a
directory name preceding them. However, if all source lives in the root
of a repository, there is no point in having a directory.

Because this check lives under 'build/include', it's impossible to
filter without filtering other useful checks.

Move it to 'build/include_directory' so it can be more easily filtered.

BUG=chromium:1073191
TEST=cros lint on a file with 'filter=-build/include_directory' in
     CPPLINT.cfg

Change-Id: Ie15fb2398a6d0297ef96219e67076ea262abf853