Update patch set 4
Patch Set 4:

These flags were there at least partially so that a tip-of-tree could continue to parse old, broken DEPS files, to aid in bisecting regressions.

This CL doesn't explain why you removed the flags, so I don't know if you were considering that need or not, but we should probably still have *some* strategy for managing bisects. 

Perhaps this should be reverted until we do? See crbug.com/809671809671#c9 for one case in point.

One alternative might be to promote an approach to bisecting where the bisect script was careful to use a contemporary version of gclient, for example, or something ...

Patch-set: 4
CC: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>