Update patch set 1
Patch Set 1:

The CROS_ROOT name makes it sound like it points to the equivalent of ~/chromiumos. Is it an existing envvar? What's the use case for this helping? How does this help with having the chroot outside the default path?

Also, what tools have you tried? I think there should be problems with a number of chromite tools when the chroot is not in the right place. That's a failing of the tools, not that it's an unreasonable thing to do (we do it on the builders after all) but we've had to be careful about things to make that work.

Patch-set: 1
Reviewer: Gerrit User 1289508 <1289508@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>