Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The CROS_ROOT name makes it sound like it points to the equivalent of ~/chromiumos. Is it an existing envvar? What's the use case for this helping? How does this help with having the chroot outside the default path?
> 
> Also, what tools have you tried? I think there should be problems with a number of chromite tools when the chroot is not in the right place. That's a failing of the tools, not that it's an unreasonable thing to do (we do it on the builders after all) but we've had to be careful about things to make that work.

Correct it should point to ~/chromiumos or equivalent so it can find chromite.  It's not an existing environment variable as far as I can tell.  cros_sdk can only be called from within the source tree right now, and it just mounts the chroot in the current directory, with no option to place it elsewhere. Only tool I really tried was cros_sdk, and then some commands inside of that to make sure things are still working. It should have no effect by default if CROS_ROOT is unset.

Patch-set: 1