Update patch set 1
Patch Set 1:

The biggest problem will be anything that has chromite.lib.path_util somewhere in its call path, it's a chroot path translation tool that currently depends on the chroot living at {SOURCE_ROOT}/chroot. If you're ok only working inside the chroot for most things, you should be able to pass through `--chroot=/whatever/path` to cros_sdk and it should largely be fine since the path_util code is short circuited when you're inside the chroot. I've built boards with that at least. It still won't work for some things outside the chroot, but at that point it should much fewer and further between, you just have to remember to always pass it to cros_sdk right now, and know what needs to be run inside and what can be run outside, which is definitely a pain.

Patch-set: 1